Add svg
