+++ src/classBased/components/TodoItem.js
@@ -1,67 +1,66 @@
-import React, { useState, useEffect } from "react"
-import styles from "./TodoItem.module.css"
+import React from 'react';
+import styles from './TodoItem.module.css';
 
-const TodoItem = props => {
-  const [editing, setEditing] = useState(false)
-
-  const handleEditing = () => {
-    setEditing(true)
+class TodoItem extends React.Component {
+  state = {
+    editing: false,
   }
-
-  const handleUpdatedDone = event => {
+  handleEditing = () => {
+    this.setState({
+      editing: true,
+    })
+  }
+  handleUpdatedDone = event => {
     if (event.key === "Enter") {
-      setEditing(false)
+      this.setState({ editing: false })
     }
   }
 
-  const completedStyle = {
-    fontStyle: "italic",
-    color: "#595959",
-    opacity: 0.4,
-    textDecoration: "line-through",
+  componentWillUnmount() {
+    console.log("Cleaning up...")
   }
 
-  const { completed, id, title } = props.todo
-
-  let viewMode = {}
-  let editMode = {}
-
-  if (editing) {
-    viewMode.display = "none"
-  } else {
-    editMode.display = "none"
-  }
-
-  useEffect(() => {
-    return () => {
-      console.log("Cleaning up...")
+  render() {
+    const completedStyle = {
+      fontStyle: "italic",
+      color: "#595959",
+      opacity: 0.4,
+      textDecoration: "line-through",
     }
-  }, [])
-
-  return (
+    const { completed, id, title } = this.props.todo
+    let viewMode = {}
+    let editMode = {}
+    
+    if (this.state.editing) {
+      viewMode.display = "none"
+    } else {
+      editMode.display = "none"
+    }
+    return (
     <li className={styles.item}>
-      <div onDoubleClick={handleEditing} style={viewMode}>
+      <div onDoubleClick={this.handleEditing} style={viewMode}>
         <input
           type="checkbox"
           className={styles.checkbox}
           checked={completed}
-          onChange={() => props.handleChangeProps(id)}
+          onChange={() => this.props.handleChangeProps(id)}
         />
-        <button onClick={() => props.deleteTodoProps(id)}>Delete</button>
-        <span style={completed ? completedStyle : null}>{title}</span>
-      </div>
-      <input
-        type="text"
-        style={editMode}
-        className={styles.textInput}
-        value={title}
-        onChange={e => {
-          props.setUpdate(e.target.value, id)
+        <button onClick={() => this.props.deleteTodoProps(id)}>Delete</button>
+         <span style={completed ? completedStyle : null}>{title}</span>
+         </div>
+         <input  
+         type="text"
+         style={editMode}
+         className={styles.textInput}
+         value={title} 
+         onChange={e => {
+          this.props.setUpdate(e.target.value, id)
         }}
-        onKeyDown={handleUpdatedDone}
-      />
-    </li>
-  )
+        onKeyDown={this.handleUpdatedDone}
+        />
+         </li>
+    );
+  }
 }
 
-export default TodoItem
+export default TodoItem;
