+++ src/classBased/components/TodosList.js
@@ -1,19 +1,22 @@
-import React from "react"
-import TodoItem from "./TodoItem"
+import React from 'react';
+import TodoItem from './TodoItem';
 
-const TodosList = props => {
-  return (
-    <ul>
-      {props.todos.map(todo => (
-        <TodoItem
-          key={todo.id}
-          todo={todo}
-          handleChangeProps={props.handleChangeProps}
-          deleteTodoProps={props.deleteTodoProps}
-          setUpdate={props.setUpdate}
-        />
-      ))}
-    </ul>
-  )
+class TodosList extends React.Component {
+  render() {
+    return (
+      <ul>
+        {this.props.todos.map((todo) => (
+          <TodoItem
+            key={todo.id}
+            todo={todo}
+            handleChangeProps={this.props.handleChangeProps}
+            deleteTodoProps={this.props.deleteTodoProps}
+            setUpdate={this.props.setUpdate}
+          />
+        ))}
+      </ul>
+    );
+  }
 }
-export default TodosList
+
+export default TodosList;
