+++ src/functionBased/components/InputTodo.js
@@ -1,4 +1,5 @@
 import React, { useState } from "react"
+import { FaPlusCircle } from "react-icons/fa"
 
 const InputTodo = props => {
   const [inputText, setInputText] = useState({
@@ -34,7 +35,11 @@ const InputTodo = props => {
         name="title"
         onChange={onChange}
       />
-      <button className="input-submit">Submit</button>
+      <button className="input-submit">
+        <FaPlusCircle
+        style={{ color: "darkcyan", fontSize: "20px", marginTop: "2px" }}
+        />
+      </button>
     </form>
   )
 }
