+++ src/functionBased/components/TodoContainer.js
@@ -1,96 +1,92 @@
-import React from 'react';
-import { v4 as uuidv4 } from 'uuid';
-import TodosList from './TodosList';
-import Header from './Header';
-import InputTodo from './InputTodo';
+import React, { useState, useEffect } from "react"
+import Header from "./Header"
+import InputTodo from "./InputTodo"
+import TodosList from "./TodosList"
+import { v4 as uuidv4 } from "uuid"
 
-class TodoContainer extends React.Component {
-    state = {
-      todos: [],
-    };
+const TodoContainer = () => {
+  function getInitialTodos() {
+    // getting stored items
+    const temp = localStorage.getItem("todos")
+    const savedTodos = JSON.parse(temp)
+    return savedTodos || []
+  }
+  const [todos, setTodos] = useState(getInitialTodos())
 
-       handleChange = (id) => {
-         this.setState((prevState) => ({
-           todos: prevState.todos.map((todo) => {
-             if (todo.id === id) {
-               return {
-                 ...todo,
-                 completed: !todo.completed,
-               };
-             }
-             return todo;
-           }),
-         }));
-       };
-
-      delTodo = (id) => {
-        this.setState({
-          todos: [
-            ...this.state.todos.filter((todo) => todo.id !== id),
-          ],
-        });
-      };
-
-      addTodoItem = (title) => {
-        const newTodo = {
-          id: uuidv4(),
-          title,
-          completed: false,
-        };
-        this.setState({
-          todos: [...this.state.todos, newTodo],
-        });
-      };
-
-      setUpdate = (updatedTitle, id) => {
-        this.setState({
-          todos: this.state.todos.map(todo => {
-            if (todo.id === id) {
-              todo.title = updatedTitle
-            }
-            return todo
-          }),
-        })
-      }
-
-      // componentDidMount() {
-      //   fetch("https://jsonplaceholder.typicode.com/todos?_limit=10")
-      //     .then(response => response.json())
-      //     .then(data => this.setState({ todos: data }));
-      // }
-
-      componentDidMount() {
-        const temp = localStorage.getItem("todos")
-        const loadedTodos = JSON.parse(temp)
-        if (loadedTodos) {
-          this.setState({
-            todos: loadedTodos
-          })
+  const handleChange = id => {
+    setTodos(prevState =>
+      prevState.map(todo => {
+        if (todo.id === id) {
+          return {
+            ...todo,
+            completed: !todo.completed,
+          }
         }
-      }
+        return todo
+      })
+    )
+  }
 
-      componentDidUpdate(prevProps, prevState) {
-        if(prevState.todos !== this.state.todos) {
-          const temp = JSON.stringify(this.state.todos)
-          localStorage.setItem("todos", temp)
+  const delTodo = id => {
+    setTodos([
+      ...todos.filter(todo => {
+        return todo.id !== id
+      }),
+    ])
+  }
+
+  const addTodoItem = title => {
+    const newTodo = {
+      id: uuidv4(),
+      title: title,
+      completed: false,
+    }
+    setTodos([...todos, newTodo])
+  }
+
+  const setUpdate = (updatedTitle, id) => {
+    setTodos(
+      todos.map(todo => {
+        if (todo.id === id) {
+          todo.title = updatedTitle
         }
-      }
+        return todo
+      })
+    )
+  }
 
-      render() {
-        return (
-          <div className="container">
-            <div className="inner">
-              <Header />
-              <InputTodo addTodoProps={this.addTodoItem} />
-              <TodosList
-                todos={this.state.todos}
-                handleChangeProps={this.handleChange}
-                deleteTodoProps={this.delTodo}
-                setUpdate={this.setUpdate}
-              />
-            </div>
-          </div>
-        );
-      }
+  useEffect(() => {
+    console.log("test run")
+  
+    // getting stored items
+    const temp = localStorage.getItem("todos")
+    const loadedTodos = JSON.parse(temp)
+  
+    if (loadedTodos) {
+      setTodos(loadedTodos)
+    }
+  }, [])
+
+  useEffect(() => {
+    // storing todos items
+    const temp = JSON.stringify(todos)
+    localStorage.setItem("todos", temp)
+  }, [todos])
+
+  return (
+    <div className="container">
+      <div className="inner">
+        <Header />
+        <InputTodo addTodoProps={addTodoItem} />
+        <TodosList
+          todos={todos}
+          handleChangeProps={handleChange}
+          deleteTodoProps={delTodo}
+          setUpdate={setUpdate}
+        />
+      </div>
+    </div>
+  )
 }
-export default TodoContainer;
+
+export default TodoContainer
