+++ src/functionBased/components/TodoItem.js
@@ -1,66 +1,70 @@
-import React from 'react';
-import styles from './TodoItem.module.css';
+import React, { useState, useEffect } from "react"
+import styles from "./TodoItem.module.css"
+import { FaTrash } from "react-icons/fa"
 
-class TodoItem extends React.Component {
-  state = {
-    editing: false,
+const TodoItem = props => {
+  const [editing, setEditing] = useState(false)
+
+  const handleEditing = () => {
+    setEditing(true)
   }
-  handleEditing = () => {
-    this.setState({
-      editing: true,
-    })
-  }
-  handleUpdatedDone = event => {
+
+  const handleUpdatedDone = event => {
     if (event.key === "Enter") {
-      this.setState({ editing: false })
+      setEditing(false)
     }
   }
 
-  componentWillUnmount() {
-    console.log("Cleaning up...")
+  const completedStyle = {
+    fontStyle: "italic",
+    color: "#595959",
+    opacity: 0.4,
+    textDecoration: "line-through",
   }
 
-  render() {
-    const completedStyle = {
-      fontStyle: "italic",
-      color: "#595959",
-      opacity: 0.4,
-      textDecoration: "line-through",
+  const { completed, id, title } = props.todo
+
+  let viewMode = {}
+  let editMode = {}
+
+  if (editing) {
+    viewMode.display = "none"
+  } else {
+    editMode.display = "none"
+  }
+
+  useEffect(() => {
+    return () => {
+      console.log("Cleaning up...")
     }
-    const { completed, id, title } = this.props.todo
-    let viewMode = {}
-    let editMode = {}
-    
-    if (this.state.editing) {
-      viewMode.display = "none"
-    } else {
-      editMode.display = "none"
-    }
-    return (
+  }, [])
+
+  return (
     <li className={styles.item}>
-      <div onDoubleClick={this.handleEditing} style={viewMode}>
+      <div onDoubleClick={handleEditing} style={viewMode}>
         <input
           type="checkbox"
           className={styles.checkbox}
           checked={completed}
-          onChange={() => this.props.handleChangeProps(id)}
+          onChange={() => props.handleChangeProps(id)}
         />
-        <button onClick={() => this.props.deleteTodoProps(id)}>Delete</button>
-         <span style={completed ? completedStyle : null}>{title}</span>
-         </div>
-         <input  
-         type="text"
-         style={editMode}
-         className={styles.textInput}
-         value={title} 
-         onChange={e => {
-          this.props.setUpdate(e.target.value, id)
+        <button onClick={() => props.deleteTodoProps(id)}>
+          <FaTrash style={{ color: "orangered", fontSize: "16px" }} />
+        </button>
+        <span style={completed ? completedStyle : null}>{title}</span>
+      </div>
+      <input
+        type="text"
+        style={editMode}
+        className={styles.textInput}
+        value={title}
+        onChange={e => {
+          props.setUpdate(e.target.value, id)
         }}
-        onKeyDown={this.handleUpdatedDone}
-        />
-         </li>
-    );
-  }
+        onKeyDown={handleUpdatedDone}
+      />
+    </li>
+  )
 }
 
-export default TodoItem;
+export default TodoItem
