+++ src/functionBased/components/TodoItem.module.css
@@ -27,3 +27,9 @@
     padding: 10px;
     border: 1px solid #dfdfdf;
   }
+
+  .input-submit svg {
+    color: darkcyan;
+    font-size: 20px;
+    margin-top: 2px;
+  }
